Resolve merge
BG commands appear to be working. What's happening is that the prompt is
being printed, the command is finishing and outputting, and then the
shell waits for input. All of this is technically correct, but makes it
look like it's hanging. Example (| is cursor):

sish:> ls &
sish:> one.txt  two.txt  ...
|